+++ backend/sooksook/src/main/java/com/smwuis/sooksook/service/study/StudyCommentService.java
@@ -1,4 +1,4 @@
-package com.smwuis.sooksook.service;
+package com.smwuis.sooksook.service.study;
 
 import com.smwuis.sooksook.domain.study.*;
 import com.smwuis.sooksook.domain.user.User;
@@ -47,20 +47,37 @@ public class StudyCommentService {
     
     // 댓글 수정
     @Transactional
-    public Long update(Long id, String content) {
+    public String update(Long id, String email, String content) {
+        User user = userRepository.findByEmail(email).orElseThrow(()-> new IllegalArgumentException("해당 유저가 없습니다."));
         StudyComment studyComment = studyCommentRepository.findById(id).orElseThrow(()-> new IllegalArgumentException("해당 댓글이 없습니다."));
-        studyComment.update(content);
-        return id;
+
+        if(user.equals(studyComment.getUserId())) {
+            studyComment.update(content);
+            return "댓글 수정 완료";
+        }
+
+        else {
+            return "댓글 수정 실패";
+        }
     }
 
     // 댓글 삭제
     @Transactional
-    public void delete(Long id) {
+    public String delete(Long id, String email) {
+        User user = userRepository.findByEmail(email).orElseThrow(()-> new IllegalArgumentException("해당 유저가 없습니다."));
         StudyComment studyComment = studyCommentRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 댓글이 없습니다."));
-        studyComment.remove();
 
-        List<StudyComment> removableCommentList = findRemovableList(id);
-        studyCommentRepository.deleteAll(removableCommentList);
+        if(user.equals(studyComment.getUserId())) {
+            studyComment.remove();
+
+            List<StudyComment> removableCommentList = findRemovableList(id);
+            studyCommentRepository.deleteAll(removableCommentList);
+            return "댓글 삭제 완료";
+        }
+
+        else {
+            return "댓글 삭제 실패";
+        }
     }
 
     // 댓글 전체 조회
